+++ games_services_platform_interface/lib/errors.dart
@@ -1,0 +1,1 @@
+export 'src/errors/leaderboard_scores_error_code.dart';